* Merged MIXER CROP, MIXER ANCHOR, MIXER ROTATION and MIXER PERSPECTIVE from 2.0 * Merged reporting of current MIXER values when no parameters are given. * Changed so that each vertex / texture coordinate pair in frame_geometry is encapsulated in a struct, to not require producers to keep track of indexes in an a array. * Changed so that the same code in image_kernel draws a single QUAD and a list of QUAD's * TODO make perspective correction and cropping work with non default geometry like the text producer. Currently no cropping or perspective correction works for custom geometry. * TODO make interaction coordinate translations work with new transforms. * Reverted text_producer change so that the baseline of the font is the actual position. * Fixed so that JSON encoded template data is also forwarded as is, instead of being interpreted as an .ftd file name. * Implemented anchor, crop, perspective and rotation support in scene_producer
Former-commit-id: 5cb5fd659f4dc4996a4be9de8a9e299e3a6d3789

+++ core/producer/text/utils/text_info.h
@@ -10,8 +10,8 @@ namespace caspar { namespace core { namespace text {
 		std::wstring font;
 		std::wstring font_file;
 
-		float					size				= 0.0f;
-		text::color<float>		color;
+		double					size				= 0.0;
+		text::color<double>		color;
 		//int					shadow_distance;
 		//int					shadow_size;
 		//float					shadow_spread;
